requirements.txt
streamlit
google-generativeai
chromadb
pandas
plotly
scikit-learn
python-dotenv
